Moved some patches/functions between Interface.cpp and MiscPatches.cpp
* And Graphics.cpp and InputFuncs/Input.cpp

Tweaked the buffer flush (actually print to file) rate in Console.cpp.
