+++ sfall/InputFuncs.h
@@ -46,8 +46,6 @@ void GetMouse(int* x, int* y);
 
 void MouseDeviceUnacquire(bool mode);
 
-void __stdcall ForceGraphicsRefresh(DWORD);
-
 #define DIK_ESCAPE          0x01
 #define DIK_1               0x02
 #define DIK_2               0x03
