+++ sfall/Modules/Interface.cpp
@@ -21,6 +21,7 @@
 #include "..\main.h"
 #include "..\FalloutEngine\Fallout2.h"
 #include "..\FalloutEngine\EngineUtils.h"
+#include "..\SimplePatch.h"
 #include "..\Utils.h"
 #include "Graphics.h"
 #include "LoadGameHook.h"
@@ -936,14 +937,6 @@ static void __declspec(naked) gmouse_bk_process_hook() {
 	}
 }
 
-static void __declspec(naked) main_death_scene_hook() {
-	__asm {
-		mov  eax, 101;
-		call fo::funcoffs::text_font_;
-		jmp  fo::funcoffs::debug_printf_;
-	}
-}
-
 static void __declspec(naked) display_body_hook() {
 	__asm {
 		mov  ebx, [esp + 0x60 - 0x28 + 8];
@@ -962,7 +955,71 @@ fix:
 	}
 }
 
+static void InterfaceWindowPatch() {
+	dlog("Applying flags patch for interface windows.", DL_INIT);
+
+	// Remove MoveOnTop flag for interfaces
+	SafeWrite8(0x46ECE9, (*(BYTE*)0x46ECE9) ^ fo::WinFlags::MoveOnTop); // Player Inventory/Loot/UseOn
+	SafeWrite8(0x41B966, (*(BYTE*)0x41B966) ^ fo::WinFlags::MoveOnTop); // Automap
+
+	// Set OwnerFlag flag
+	SafeWrite8(0x4D5EBF, fo::WinFlags::OwnerFlag); // win_init_ (main win)
+	SafeWrite8(0x481CEC, (*(BYTE*)0x481CEC) | fo::WinFlags::OwnerFlag); // _display_win (map win)
+	SafeWrite8(0x44E7D2, (*(BYTE*)0x44E7D2) | fo::WinFlags::OwnerFlag); // gmovie_play_ (movie win)
+
+	// Remove OwnerFlag flag
+	SafeWrite8(0x4B801B, (*(BYTE*)0x4B801B) ^ fo::WinFlags::OwnerFlag); // createWindow_
+	// Remove OwnerFlag and Transparent flags
+	SafeWrite8(0x42F869, (*(BYTE*)0x42F869) ^ (fo::WinFlags::Transparent | fo::WinFlags::OwnerFlag)); // addWindow_
+
+	dlogr(" Done", DL_INIT);
+
+	// Cosmetic fix for the background image of the character portrait on the player's inventory screen
+	HookCall(0x47093C, display_body_hook);
+	BYTE code[11] = {
+		0x8B, 0xD3,             // mov  edx, ebx
+		0x66, 0x8B, 0x58, 0xF4, // mov  bx, [eax - 12] [sizeof(frame)]
+		0x0F, 0xAF, 0xD3,       // imul edx, ebx (y * frame width)
+		0x53, 0x90              // push ebx (frame width)
+	};
+	SafeWriteBytes(0x470971, code, 11); // calculates the offset in the pixel array for x/y coordinates
+
+	// Increase the max text width of the player name on the character screen
+	SafeWriteBatch<BYTE>(127, {0x435160, 0x435189}); // 100
+
+	// Increase the max text width of the information card on the character screen
+	SafeWriteBatch<BYTE>(145, {0x43ACD5, 0x43DD37}); // 136, 133
+}
+
+static void InventoryCharacterRotationSpeedPatch() {
+	long setting = IniReader::GetConfigInt("Misc", "SpeedInventoryPCRotation", 166);
+	if (setting != 166 && setting <= 1000) {
+		dlog("Applying SpeedInventoryPCRotation patch.", DL_INIT);
+		SafeWrite32(0x47066B, setting);
+		dlogr(" Done", DL_INIT);
+	}
+}
+
+static void UIAnimationSpeedPatch() {
+	DWORD addrs[] = {
+		0x45F9DE, 0x45FB33,
+		0x447DF4, 0x447EB6,
+		0x499B99, 0x499DA8
+	};
+	SimplePatch<WORD>(addrs, 2, "Misc", "CombatPanelAnimDelay", 1000, 0, 65535);
+	SimplePatch<BYTE>(&addrs[2], 2, "Misc", "DialogPanelAnimDelay", 33, 0, 255);
+	SimplePatch<BYTE>(&addrs[4], 2, "Misc", "PipboyTimeAnimDelay", 50, 0, 127);
+}
+
 void Interface::init() {
+	InterfaceWindowPatch();
+	InventoryCharacterRotationSpeedPatch();
+	UIAnimationSpeedPatch();
+
+	if (IniReader::GetConfigInt("Misc", "RemoveWindowRounding", 1)) {
+		SafeWriteBatch<BYTE>(CodeType::JumpShort, {0x4D6EDD, 0x4D6F12});
+	}
+
 	if (IniReader::GetConfigInt("Interface", "ActionPointsBar", 0)) {
 		ActionPointsBarPatch();
 	}
@@ -978,28 +1035,6 @@ void Interface::init() {
 		if (hrpVersionValid) IFACE_BAR_MODE = (GetIntHRPValue(HRP_VAR_IFACE_BAR_MODE) != 0);
 		HookCall(0x44C018, gmouse_handle_event_hook); // replaces hack function from HRP
 	};
-
-	// Set the normal font for death screen subtitles
-	if (IniReader::GetConfigInt("Misc", "DeathScreenFontPatch", 0)) {
-		dlog("Applying death screen font patch.", DL_INIT);
-		HookCall(0x4812DF, main_death_scene_hook);
-		dlogr(" Done", DL_INIT);
-	}
-
-	// Corrects the height of the black background for death screen subtitles
-	if (!hrpIsEnabled) SafeWrite32(0x48134D, 38 - (640 * 3));      // main_death_scene_ (shift y-offset 2px up, w/o HRP)
-	if (!hrpIsEnabled || hrpVersionValid) SafeWrite8(0x481345, 4); // main_death_scene_
-	if (hrpVersionValid) SafeWrite8(HRPAddress(0x10011738), 10);
-
-	// Cosmetic fix for the background image of the character portrait on the player's inventory screen
-	HookCall(0x47093C, display_body_hook);
-	BYTE code[11] = {
-		0x8B, 0xD3,             // mov  edx, ebx
-		0x66, 0x8B, 0x58, 0xF4, // mov  bx, [eax - 12] [sizeof(frame)]
-		0x0F, 0xAF, 0xD3,       // imul edx, ebx (y * frame width)
-		0x53, 0x90              // push ebx (frame width)
-	};
-	SafeWriteBytes(0x470971, code, 11); // calculates the offset in the pixel array for x/y coordinates
 }
 
 void Interface::exit() {
